refactor: marketing section
