+++ src/pages/marketing.tsx
@@ -3,7 +3,7 @@ import { OgMetaTags } from '@src/components/og-meta-tags.tsx';
 import Header from "@src/layouts/dashboard/header.tsx";
 import HeaderContent from "@src/layouts/dashboard/header-content.tsx";
 import {WithAuth} from "@src/components/should-login/withAuth.tsx";
-import StrategyOverview from "@src/pages/dashboard/marketing/components/StrategyOverview.tsx";
+import StrategyOverview from "@src/pages/marketing/components/StrategyOverview.tsx";
 
 // ----------------------------------------------------------------------
