+++ src/pages/marketing/index.tsx
@@ -3,7 +3,7 @@ import Header from '@src/layouts/dashboard/header.tsx';
 import HeaderContent from '@src/layouts/dashboard/header-content.tsx';
 import { WithAuth } from '@src/components/should-login/withAuth.tsx';
 import { OgMetaTags } from '@src/components/og-meta-tags.tsx';
-import StrategyOverview from '@src/pages/dashboard/marketing/components/StrategyOverview.tsx';
+import MarketingView from '@src/sections/marketing/views/marketing-view.tsx';
 
 const StrategyIndex = () => {
   return (
@@ -16,7 +16,7 @@ const StrategyIndex = () => {
         <HeaderContent title="Marketing" />
       </Header>
       <WithAuth
-        component={StrategyOverview}
+        component={MarketingView}
         header={'Marketing Dashboard'}
         description={'Login to access your marketing overview.'}
         icon={'iconoir:megaphone'}
