+++ src/sections/marketing/components/campaign-modal-content.tsx
@@ -19,11 +19,11 @@ import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
 import { useCreateCampaign } from '@src/hooks/protocol/use-create-campaign.ts';
 import { SelectChangeEvent } from '@mui/material/Select';
 import LoadingButton from "@mui/lab/LoadingButton";
-import NeonPaper from '@src/sections/publication/NeonPaperContainer';
+import NeonPaper from '@src/sections/publication/components/neon-paper-container.tsx';
 import { notifySuccess } from '@notifications/internal-notifications.ts';
 import { SUCCESS } from '@notifications/success.ts';
-import { NEW_CAMPAIGN_POLICIES_OPTIONS } from '@src/sections/marketing/components/CONSTANTS.tsx';
-import { CampaignModalContentProps } from '@src/sections/marketing/components/types.ts';
+import { NEW_CAMPAIGN_POLICIES_OPTIONS } from '@src/sections/marketing/CONSTANTS.tsx';
+import { CampaignModalContentProps } from '@src/sections/marketing/types.ts';
 
 // ----------------------------------------------------------------------
