+++ src/sections/marketing/components/campaign-withdraw-funds-modal-content.tsx
@@ -11,11 +11,11 @@ import { parseUnits } from 'viem';
 
 // LOCAL IMPORTS
 import { useCampaignRemoveFunds } from '@src/hooks/protocol/use-campaign-remove-funds.ts';
-import NeonPaper from "@src/sections/publication/NeonPaperContainer";
+import NeonPaper from "@src/sections/publication/components/neon-paper-container.tsx";
 import { notifyError, notifySuccess } from '@notifications/internal-notifications.ts';
 import { ERRORS } from '@notifications/errors.ts';
 import { SUCCESS } from '@notifications/success.ts';
-import { CampaignWithdrawFundsModalContentProps } from '@src/sections/marketing/components/types.ts';
+import { CampaignWithdrawFundsModalContentProps } from '@src/sections/marketing/types.ts';
 
 const CampaignWithdrawFundsModalContent: FC<CampaignWithdrawFundsModalContentProps> = (props) => {
   const { campaignData, onClose, onConfirm } = props;
