+++ src/sections/marketing/index.tsx
@@ -1,0 +1,1 @@
+export { default as MarketingView } from '@src/sections/marketing/views/marketing-view.tsx';
